+++ packages/UsageAnalysis/src/tabs/log.ts
@@ -1,5 +1,6 @@
 import * as DG from 'datagrok-api/dg';
 import * as ui from 'datagrok-api/ui';
+import * as grok from 'datagrok-api/grok';
 
 import {UaView} from './ua';
 import {UaFilterableQueryViewer} from '../viewers/ua-filterable-query-viewer';
@@ -36,6 +37,10 @@ export class LogView extends UaView {
       filterSubscription: this.uaToolbox.filterStream,
       name: 'Log',
       queryName: 'LogTail',
+      processDataFrame: (t: DG.DataFrame) => {
+        t.onCurrentRowChanged.subscribe(() => this.showLogContextPanel(t));
+        return t;
+      },
       createViewer: (t: DG.DataFrame) => {
         const viewer = DG.Viewer.grid(t, {
           'showColumnLabels': false,
@@ -81,5 +86,28 @@ export class LogView extends UaView {
       filters,
       logViewer.root,
     ]));
+  }
+
+  showLogContextPanel(table: DG.DataFrame): void {
+    const rowIdx = table.currentRowIdx;
+    if (rowIdx < 0)
+      return;
+    const eventId = table.getCol('id').get(rowIdx);
+    if (!eventId)
+      return;
+    const accordion = DG.Accordion.create();
+    accordion.addPane('Details', () => ui.wait(async () => {
+      const t: DG.DataFrame = await grok.functions.call('UsageAnalysis:LogEventParameters', {eventId});
+      if (t.rowCount === 0)
+        return ui.divText('No details available');
+      const names = t.getCol('param_name').toList();
+      const values = t.getCol('value').toList();
+      const map: {[key: string]: string} = {};
+      for (let i = 0; i < names.length; i++)
+        map[names[i]] = values[i];
+      return ui.tableFromMap(map);
+    }), true);
+
+    grok.shell.o = accordion.root;
   }
 }